Moving fragments to separate folder
