+++ app/src/main/java/com/example/eddy/servr/activities/MainActivity.java
@@ -14,6 +14,7 @@ import android.view.MenuInflater;
 import android.view.MenuItem;
 
 import com.example.eddy.servr.R;
+import com.example.eddy.servr.fragments.*;
 
 import java.util.ArrayList;
 import java.util.List;
