+++ app/src/main/java/com/example/eddy/servr/fragments/Fragment_3.java
@@ -1,4 +1,4 @@
-package com.example.eddy.servr.activities;
+package com.example.eddy.servr.fragments;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -8,12 +8,12 @@ import android.view.ViewGroup;
 
 import com.example.eddy.servr.R;
 
-public class Fragment_1 extends Fragment{
+public class Fragment_3 extends Fragment{
 
-    public Fragment_1(){}
+    public Fragment_3(){}
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_1, container, false);
+        return inflater.inflate(R.layout.fragment_3, container, false);
     }
 }
